rns: extract RNSGatewayManager, move dialling off the announce thread (phase 7)
Advertise-host resolution, self-skip, endpoint cooldown, the interface cap and
the dynamic BackboneClientInterface registration move into RNSGatewayManager,
along with forceBackboneReconnect() (the circuit breaker's counter stays in the
loop that trips it).

Behaviour change: the blocking part of a dynamic gateway add now runs on a
dedicated single-thread executor ("RNS-GatewayDial", bounded queue 8, discard
policy) instead of inline. It previously ran on Reticulum's announce-delivery
thread, inside QAnnounceHandler.receivedAnnounce — constructing a
BackboneClientInterface, running initIFac() and calling launch(), i.e. a TCP
connect, so one slow or unreachable announced gateway stalled announce
processing for every peer. All the cheap checks (usable host, self-skip,
cooldown stamping, existing-interface dedup, initiator cap) still run inline so
ordering and dedup stay deterministic; only the launch is handed off. A
discarded dial retries after the cooldown anyway.

@Synchronized on receivedAnnounce is kept, with a comment on what it does and
does not cover — it does not serialise BASE against DATA, since the two handler
instances hold separate Lombok $locks. The analysis flagged it because of the
network I/O it was holding across; that is now gone, and the lock itself is
cheap, so removing it would be a concurrency change with no benefit.

Also cleaned up the fully-qualified names left inline where the import already
existed (ConcurrentHashMap, Future, RejectedExecutionException, TimeUnit,
ConnectionInterface, Duration, HashMap, Hex.decodeHex).

RNS.java: 1876 -> 1701 lines.
